+++ test/identify/index.spec.ts
@@ -29,6 +29,7 @@ import pDefer from 'p-defer'
 import { defaultComponents, Components } from '../../src/components.js'
 import { stubInterface } from 'sinon-ts'
 import type { TransportManager } from '@libp2p/interface-transport'
+import { DefaultIdentifyService } from '../../src/identify/identify.js'
 
 const listenMaddrs = [multiaddr('/ip4/127.0.0.1/tcp/15002/ws')]
 
@@ -101,8 +102,8 @@ describe('identify', () => {
   })
 
   it('should be able to identify another peer', async () => {
-    const localIdentify = identifyService(defaultInit)(localComponents)
-    const remoteIdentify = identifyService(defaultInit)(remoteComponents)
+    const localIdentify = new DefaultIdentifyService(localComponents, defaultInit)
+    const remoteIdentify = new DefaultIdentifyService(remoteComponents, defaultInit)
 
     await start(localIdentify)
     await start(remoteIdentify)
@@ -124,8 +125,8 @@ describe('identify', () => {
   })
 
   it('should throw if identified peer is the wrong peer', async () => {
-    const localIdentify = identifyService(defaultInit)(localComponents)
-    const remoteIdentify = identifyService(defaultInit)(remoteComponents)
+    const localIdentify = new DefaultIdentifyService(localComponents, defaultInit)
+    const remoteIdentify = new DefaultIdentifyService(remoteComponents, defaultInit)
 
     await start(localIdentify)
     await start(remoteIdentify)
@@ -187,8 +188,8 @@ describe('identify', () => {
   })
 
   it('should time out during identify', async () => {
-    const localIdentify = identifyService(defaultInit)(localComponents)
-    const remoteIdentify = identifyService(defaultInit)(remoteComponents)
+    const localIdentify = new DefaultIdentifyService(localComponents, defaultInit)
+    const remoteIdentify = new DefaultIdentifyService(remoteComponents, defaultInit)
 
     await start(localIdentify)
     await start(remoteIdentify)
@@ -233,10 +234,10 @@ describe('identify', () => {
   it('should limit incoming identify message sizes', async () => {
     const deferred = pDefer()
 
-    const remoteIdentify = identifyService({
+    const remoteIdentify = new DefaultIdentifyService(remoteComponents, {
       ...defaultInit,
       maxIdentifyMessageSize: 100
-    })(remoteComponents)
+    })
     await start(remoteIdentify)
 
     const identifySpy = sinon.spy(remoteIdentify, 'identify')
@@ -279,10 +280,10 @@ describe('identify', () => {
   it('should time out incoming identify messages', async () => {
     const deferred = pDefer()
 
-    const remoteIdentify = identifyService({
+    const remoteIdentify = new DefaultIdentifyService(remoteComponents, {
       ...defaultInit,
       timeout: 100
-    })(remoteComponents)
+    })
     await start(remoteIdentify)
 
     const identifySpy = sinon.spy(remoteIdentify, 'identify')
